DI: lift Context out of EL
