+++ sig/datadog/di/probe_manager.rbs
@@ -57,7 +57,7 @@ module Datadog
 
       public
       def install_pending_line_probes: (untyped path) -> untyped
-      def probe_executed_callback: (EL::Context context) -> untyped
+      def probe_executed_callback: (Context context) -> untyped
       attr_reader definition_trace_point: untyped
     end
   end
