+++ spec/datadog/di/instrumenter_spec.rb
@@ -89,7 +89,7 @@ RSpec.describe Datadog::DI::Instrumenter do
         expect(HookTestClass.new.hook_test_method).to eq 42
 
         expect(observed_calls.length).to eq 1
-        expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+        expect(observed_calls.first).to be_a(Datadog::DI::Context)
         expect(observed_calls.first.return_value).to eq 42
         expect(observed_calls.first.duration).to be_a(Float)
       end
@@ -115,7 +115,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             expect(yielded_value).to eq([['hello'], {}])
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq [['hello'], {}]
             expect(observed_calls.first.duration).to be_a(Float)
           end
@@ -162,7 +162,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             expect(yielded_value).to eq(expected_rv)
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq expected_rv
             expect(observed_calls.first.duration).to be_a(Float)
           end
@@ -205,7 +205,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             expect(yielded_value).to eq([['hello'], {kw: 'world'}])
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq [['hello'], {kw: 'world'}]
             expect(observed_calls.first.duration).to be_a(Float)
           end
@@ -248,7 +248,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             expect(yielded_value).to eq([['hello'], {kw: 'world'}])
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq [['hello'], {kw: 'world'}]
             expect(observed_calls.first.duration).to be_a(Float)
           end
@@ -317,7 +317,7 @@ RSpec.describe Datadog::DI::Instrumenter do
         target_call
 
         expect(observed_calls.length).to eq 1
-        expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+        expect(observed_calls.first).to be_a(Datadog::DI::Context)
         expect(observed_calls.first.return_value).to eq 42
         expect(observed_calls.first.duration).to be_a(Float)
 
@@ -346,7 +346,7 @@ RSpec.describe Datadog::DI::Instrumenter do
           expect(HookTestClass.new.hook_test_method_with_arg(2)).to eq 2
 
           expect(observed_calls.length).to eq 1
-          expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls.first).to be_a(Datadog::DI::Context)
           expect(observed_calls.first.return_value).to eq 2
           expect(observed_calls.first.duration).to be_a(Float)
           # expect(observed_calls.first.serialized_entry_args).to eq(arg1: 2)
@@ -372,7 +372,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             target_call
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq 2
             expect(observed_calls.first.duration).to be_a(Float)
 
@@ -400,7 +400,7 @@ RSpec.describe Datadog::DI::Instrumenter do
               target_call
 
               expect(observed_calls.length).to eq 1
-              expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+              expect(observed_calls.first).to be_a(Datadog::DI::Context)
               expect(observed_calls.first.return_value).to eq 2
               expect(observed_calls.first.duration).to be_a(Float)
 
@@ -462,7 +462,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             target_call
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq 42
             expect(observed_calls.first.duration).to be_a(Float)
 
@@ -514,7 +514,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             target_call
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq 42
             expect(observed_calls.first.duration).to be_a(Float)
 
@@ -552,7 +552,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             target_call
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq [41, 42]
             expect(observed_calls.first.duration).to be_a(Float)
 
@@ -604,7 +604,7 @@ RSpec.describe Datadog::DI::Instrumenter do
           target_call
 
           expect(observed_calls.length).to eq 1
-          expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls.first).to be_a(Datadog::DI::Context)
           expect(observed_calls.first.return_value).to eq(kwarg: 42)
           expect(observed_calls.first.duration).to be_a(Float)
 
@@ -659,7 +659,7 @@ RSpec.describe Datadog::DI::Instrumenter do
             target_call
 
             expect(observed_calls.length).to eq 1
-            expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls.first).to be_a(Datadog::DI::Context)
             expect(observed_calls.first.return_value).to eq(['hello', {kwarg: 42}])
             expect(observed_calls.first.duration).to be_a(Float)
 
@@ -722,7 +722,7 @@ RSpec.describe Datadog::DI::Instrumenter do
         end.to raise_error(HookTestClass::TestException)
 
         expect(observed_calls.length).to eq 1
-        expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+        expect(observed_calls.first).to be_a(Datadog::DI::Context)
         expect(observed_calls.first.return_value).to be nil
         expect(observed_calls.first.exception).to be_a(HookTestClass::TestException)
         expect(observed_calls.first.duration).to be_a(Float)
@@ -765,7 +765,7 @@ RSpec.describe Datadog::DI::Instrumenter do
         expect(HookTestClass.new.hook_test_method).to eq 42
 
         expect(observed_calls.length).to eq 2
-        expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+        expect(observed_calls.first).to be_a(Datadog::DI::Context)
         expect(observed_calls.first.return_value).to eq 42
         expect(observed_calls.first.duration).to be_a(Float)
 
@@ -853,19 +853,19 @@ RSpec.describe Datadog::DI::Instrumenter do
 
           # TODO add assertions for parameters and locals
 
-          expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls.first).to be_a(Datadog::DI::Context)
           expect(observed_calls[0].return_value).to eq '+'
           expect(observed_calls[0].duration).to be_a(Float)
 
-          expect(observed_calls[1]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[1]).to be_a(Datadog::DI::Context)
           expect(observed_calls[1].return_value).to eq '+-'
           expect(observed_calls[1].duration).to be_a(Float)
 
-          expect(observed_calls[2]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[2]).to be_a(Datadog::DI::Context)
           expect(observed_calls[2].return_value).to eq '+--'
           expect(observed_calls[2].duration).to be_a(Float)
 
-          expect(observed_calls[3]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[3]).to be_a(Datadog::DI::Context)
           expect(observed_calls[3].return_value).to eq '+---'
           expect(observed_calls[3].duration).to be_a(Float)
         end
@@ -1019,8 +1019,8 @@ RSpec.describe Datadog::DI::Instrumenter do
 
           expect(observed_calls.length).to eq 2
 
-          expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
-          expect(observed_calls[1]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls.first).to be_a(Datadog::DI::Context)
+          expect(observed_calls[1]).to be_a(Datadog::DI::Context)
         end
       end
     end
@@ -1164,7 +1164,7 @@ RSpec.describe Datadog::DI::Instrumenter do
       end
 
       it 'invokes callback with expected keys' do
-        expect(payload).to be_a(Datadog::DI::EL::Context)
+        expect(payload).to be_a(Datadog::DI::Context)
       end
 
       describe 'stack trace' do
@@ -1216,11 +1216,11 @@ RSpec.describe Datadog::DI::Instrumenter do
         HookLineBasicTestClass.new.test_method
 
         expect(observed_calls.length).to eq 2
-        expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+        expect(observed_calls.first).to be_a(Datadog::DI::Context)
         # We do not have locals here because we are not capturing,
         # but we do have path which came from the trace point object.
         expect(observed_calls.first.path).to be_a(String)
-        expect(observed_calls[1]).to be_a(Datadog::DI::EL::Context)
+        expect(observed_calls[1]).to be_a(Datadog::DI::Context)
         expect(observed_calls[1].path).to be_a(String)
       end
     end
@@ -1253,7 +1253,7 @@ RSpec.describe Datadog::DI::Instrumenter do
         HookLineTargetedTestClass.new.test_method
 
         expect(observed_calls.length).to eq 1
-        expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+        expect(observed_calls.first).to be_a(Datadog::DI::Context)
       end
 
       context 'end line of a method' do
@@ -1277,7 +1277,7 @@ RSpec.describe Datadog::DI::Instrumenter do
 
           expect(observed_calls.length).to eq 1
 
-          expect(observed_calls.first).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls.first).to be_a(Datadog::DI::Context)
         end
 
         # Since the instrumentation mutates the state of the probe,
@@ -1295,8 +1295,8 @@ RSpec.describe Datadog::DI::Instrumenter do
 
             expect(observed_calls.length).to eq 2
 
-            expect(observed_calls[0]).to be_a(Datadog::DI::EL::Context)
-            expect(observed_calls[1]).to be_a(Datadog::DI::EL::Context)
+            expect(observed_calls[0]).to be_a(Datadog::DI::Context)
+            expect(observed_calls[1]).to be_a(Datadog::DI::Context)
           end
         end
       end
@@ -1341,16 +1341,16 @@ RSpec.describe Datadog::DI::Instrumenter do
 
           # TODO add assertions for locals
 
-          expect(observed_calls[0]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[0]).to be_a(Datadog::DI::Context)
           expect(observed_calls[0].caller_locations).to be_a(Array)
 
-          expect(observed_calls[1]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[1]).to be_a(Datadog::DI::Context)
           expect(observed_calls[1].caller_locations).to be_a(Array)
 
-          expect(observed_calls[2]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[2]).to be_a(Datadog::DI::Context)
           expect(observed_calls[2].caller_locations).to be_a(Array)
 
-          expect(observed_calls[3]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[3]).to be_a(Datadog::DI::Context)
           expect(observed_calls[3].caller_locations).to be_a(Array)
         end
       end
@@ -1389,7 +1389,7 @@ RSpec.describe Datadog::DI::Instrumenter do
           # generating one snapshot.
           expect(observed_calls.length).to eq 1
 
-          expect(observed_calls[0]).to be_a(Datadog::DI::EL::Context)
+          expect(observed_calls[0]).to be_a(Datadog::DI::Context)
           expect(observed_calls[0].caller_locations).to be_a(Array)
         end
       end
@@ -1423,7 +1423,7 @@ RSpec.describe Datadog::DI::Instrumenter do
       end
 
       it 'invokes callback with expected keys' do
-        expect(payload).to be_a(Datadog::DI::EL::Context)
+        expect(payload).to be_a(Datadog::DI::Context)
       end
     end
